% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sleep.R
\name{sleep}
\alias{sleep}
\title{Sleep in seconds}
\usage{
sleep(seconds)
}
\arguments{
\item{seconds}{time in seconds. can be fractional. will be converted to
microseconds and called with 'usleep()' in C}
}
\description{
Avoid \code{Sys.sleep()} in event loops as per docuemntation for
\code{grDevices:::getGraphicsEvent}.  \code{Sys.sleep()} actually removes
pending graphics events and is going to mess up this whole system.
}
