% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-governor.R
\name{fps_governor}
\alias{fps_governor}
\title{Governor for frame rate}
\usage{
fps_governor(fps)
}
\arguments{
\item{fps}{target frames per second}
}
\value{
actual fps calculated over last 10 frames
}
\description{
Given a target FPS, this function will insert pauses to ensure it can
be called no more than this number of times per second (give or take a little
bit)
}
