% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-loop.R
\name{run_loop}
\alias{run_loop}
\title{Run the user supplied function as the idle function within the event loop}
\usage{
run_loop(user_func, width = 7, height = 7, target_fps = 30)
}
\arguments{
\item{user_func}{user function}

\item{width, height}{size of graphics device to open. Default: 7x7 inches}

\item{target_fps}{target frames-per-second.  If rendering speed surpasses
this then slight pauses will be added to each loop to bring this
back to the target rate}
}
\value{
NULL.  The user function is run over-and-over within the event
        loop.
}
\description{
Run the user supplied function as the idle function within the event loop
}
