% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sleep.R
\name{sleep}
\alias{sleep}
\title{Sleep in seconds}
\usage{
sleep(seconds)
}
\arguments{
\item{seconds}{time in seconds. can be fractional. will be converted to
microseconds and called with 'usleep()' in C}
}
\description{
This is a custom bit of C code to pause the computer for a bit.
}
\details{
For an eventloop, need to avoid \code{Sys.sleep()}, as discussed in the
 documentation for
\code{grDevices:::getGraphicsEvent}.  This is because \code{Sys.sleep()} actually removes
pending graphics events and is going to mess up this whole system.
}
